refactor: type분리
