+++ src/components/Articles/OfferItemLists/index.tsx
@@ -1,21 +1,9 @@
 import styled from 'styled-components';
 import OfferItem from '@components/Articles/OfferItem';
-
-export interface OfferTypes {
-  profileImg: string;
-  nickname: string;
-  location: string;
-  rating: string;
-  title: string;
-  category: string;
-  price: string;
-  createdAt: Date | string;
-  text: string;
-  isOriginalPost: boolean;
-}
+import { OfferPostTypes } from '@/types/OfferTypes';
 
 interface IOfferItemListsProps {
-  offers: OfferTypes[];
+  offers: OfferPostTypes[];
 }
 
 const Ul = styled.ul``;
