+++ src/components/SignUp/KakaoMap/index.tsx
@@ -3,7 +3,7 @@ import * as K from './KakaoMapStyles';
 import useCurrentPosition from '@/hooks/useCurrentPosition';
 import CurrentLocation from '@/components/SignUp/CurrentLocation';
 import SetCurrentLocationBtn from '@/components/SignUp/SetCurrentLocationBtn';
-import { Coordinates } from '@Pages/SingUp';
+import { Coordinates } from '@/types/UserTypes';
 
 interface IKakaoMapProps {
   selectedAddress: string;
