+++ src/types/OfferTypes.ts
@@ -1,0 +1,12 @@
+export interface OfferPostTypes {
+  profileImg: string;
+  nickname: string;
+  location: string;
+  rating: string;
+  title: string;
+  category: string;
+  price: string;
+  createdAt: Date | string;
+  text: string;
+  isOriginalPost: boolean;
+}
